requirements.txt
langchain>=0.1.12
langchain-community>=0.0.28
langchain-openai>=0.0.5
openai==1.14.0
python-dotenv==1.0.1
pypdf==4.1.0
yt-dlp==2024.3.10
pydub==0.25.1
chromadb==0.4.22
numpy>=1.24.0
tiktoken>=0.5.2
sentence-transformers>=2.2.2
urllib3<2.0.0